requirements.txt
pandas
numpy
scikit-learn
scipy
Flask
pyYAML
ensure
requests
nltk
tensorflow
-e .